backend/requirements.txt
asgiref==3.6.0
attrs==22.2.0
Automat==22.10.0
black==22.10.0
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
cssselect==1.2.0
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.6
Django==4.1.4
django-cors-headers==3.13.0
django-import-export==3.1.0
django-rest-passwordreset==1.3.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
et-xmlfile==1.1.0
filelock==3.9.1
flake8==4.0.1
flake8-annotations==2.9.1
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
hyperlink==21.0.0
identify==2.5.20
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
itemadapter==0.7.0
itemloaders==1.0.6
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.2
MarkupPy==1.14
MarkupSafe==2.1.2
mccabe==0.6.1
mypy-extensions==1.0.0
nodeenv==1.7.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.0
parsel==1.7.0
pathspec==0.11.0
pep8-naming==0.13.2
platformdirs==2.6.2
pluggy==1.0.0
pre-commit==2.20.0
Protego==0.2.1
psycopg2-binary==2.9.5
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyDispatcher==2.0.7
pyflakes==2.4.0
PyJWT==2.6.0
pyOpenSSL==23.0.0
pytest==6.2.5
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
queuelib==1.6.2
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
Scrapy==2.8.0
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tablib==3.3.0
tldextract==3.4.0
toml==0.10.2
tomli==2.0.1
Twisted==22.10.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.17.1
w3lib==2.1.1
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.5.2
